+++ hmda/src/main/scala/hmda/api/http/public/LarValidationHttpApi.scala
@@ -3,17 +3,11 @@ package hmda.api.http.public
 import akka.actor.ActorSystem
 import akka.event.LoggingAdapter
 import akka.http.scaladsl.marshalling.ToResponseMarshallable
-import akka.http.scaladsl.model.StatusCodes
 import akka.stream.ActorMaterializer
 import akka.util.Timeout
 import akka.http.scaladsl.server.Directives._
 import akka.http.scaladsl.server.Route
-import hmda.api.http.model.public.{
-  LarValidateRequest,
-  LarValidateResponse,
-  SingleValidationErrorResult,
-  ValidationErrorSummary
-}
+import hmda.api.http.model.public.LarValidateRequest
 import hmda.parser.filing.lar.LarCsvParser
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport._
 import io.circe.generic.auto._
@@ -21,13 +15,13 @@ import hmda.api.http.codec.filing.LarCodec._
 import hmda.api.http.directives.HmdaTimeDirectives
 import ch.megard.akka.http.cors.scaladsl.CorsDirectives._
 import hmda.model.filing.lar.LoanApplicationRegister
-import hmda.model.validation._
-import hmda.parser.ParserErrorModel
 import hmda.validation.engine.LarEngine._
 
 import scala.concurrent.ExecutionContext
 
-trait LarValidationHttpApi extends HmdaTimeDirectives {
+trait LarValidationHttpApi
+    extends HmdaTimeDirectives
+    with FilingValidationHttpApi {
 
   implicit val system: ActorSystem
   implicit val materializer: ActorMaterializer
@@ -69,14 +63,6 @@ trait LarValidationHttpApi extends HmdaTimeDirectives {
       }
     }
 
-  private def completeWithParsingErrors(
-      errors: List[ParserErrorModel.ParserValidationError]): Route = {
-    val errorList = errors.map(e => e.errorMessage)
-    complete(
-      ToResponseMarshallable(
-        StatusCodes.BadRequest -> LarValidateResponse(errorList)))
-  }
-
   private def validate(lar: LoanApplicationRegister,
                        checkType: String): Route = {
     val validation: HmdaValidation[LoanApplicationRegister] = checkType match {
@@ -91,23 +77,8 @@ trait LarValidationHttpApi extends HmdaTimeDirectives {
     maybeErrors match {
       case Right(l) => complete(l)
       case Left(errors) =>
-        complete(ToResponseMarshallable(aggregateLarErrors(errors)))
+        complete(ToResponseMarshallable(aggregateErrors(errors)))
     }
-  }
-
-  private def aggregateLarErrors(
-      errors: List[ValidationError]): SingleValidationErrorResult = {
-    val groupedErrors = errors.groupBy(_.validationErrorType)
-    def allOfType(errorType: ValidationErrorType): Seq[String] = {
-      groupedErrors.getOrElse(errorType, List()).map(e => e.editName)
-    }
-
-    SingleValidationErrorResult(
-      ValidationErrorSummary(allOfType(Syntactical)),
-      ValidationErrorSummary(allOfType(Validity)),
-      ValidationErrorSummary(allOfType(Quality))
-    )
-
   }
 
   def larRoutes: Route = {
